server: detach emit_certificate_req() from ExpectClientHello
